# Skip annotation of vendor chunks for performance reasons