--- a/package.json
+++ b/package.json
@@ -1,6 +1,6 @@
 {
   "name": "ng-annotate-webpack-plugin",
-  "version": "0.3.0",
+  "version": "0.4.0",
   "description": "webpack plugin that runs ng-annotate on your bundles",
   "main": "index.js",
   "scripts": {