Using `ng-annotate-webpack-plugin` can slow down builds a lot, especially in development mode. In a small playground project with AngularJS 1.7.8 and Angular 8, I get the following build times before and after the changes from this PR:

Production build (`webpack -p`, no sourcemaps, AOT): before 8.9 s, after: 6.2 s
Development build (`webpack-dev-server`, sourcemaps, JIT): **before: 64 s** (!) after: 6.9 s

It looks like `ng-annotate` is slowing down the development build by almost a factor of 10 when run on the vendor chunk.

This pull request disables annotation of vendor chunks and exposes an option to customize the behavior.