@jeffling 2019-10-14
Would be happy to merge your changes in, though the internal references now point to your fork ( as well as the README ) which may affect existing dependants. if you update the references to your fork I could go ahead and merge. 

Thanks! 

Note: You may want to look into https://github.com/schmod/babel-plugin-angularjs-annotate which is probably the right way to do angular 1 annotation.