@rhuitl 2019-10-15
Yes, that changed was not intended to end up in this pull request. Created #29 instead.

Thanks for the pointer! We're using TypeScript and running its output through Babel only to get annotations doesn't make too much sense in this case, I guess.